[actions] don't run Amplify on PRs from forks
GitHub does not mint OIDC tokens for externally sourced PRs so this
workflow can't successfully run. An alternative solution (like via an
approval comment?) should be identified and implemented eventually to
allow the workflow for previous contributors using their own forks.

This also updates the workflow to run on the develop branch.